requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-yasg==1.21.3
factory-boy==3.2.1
Faker==13.15.1
flake8==5.0.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.0
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
